splitting information printing and bracket-trader finding into two scripts
